requirements.txt
google_api_python_client==2.83.0
google_auth_oauthlib==0.4.1
protobuf==4.25.2
streamlit==1.31.0
tiktoken==0.5.2
unstructured==0.11.8
google-generativeai==0.3.2
openai==1.11.1
